assistant/requirements.txt
vosk
piper-tts
llama-cpp-python
rapidfuzz
PyYAML
scikit-learn
joblib
